requirements/requirements.txt
poetry-core>=1.6.1
--trusted-host nexus3.00.egov.local
--extra-index-url https://nexus3.00.egov.local/repository/da-group/simple

aiobotocore==2.5.1 ; python_full_version == "3.9.13"
aiobotocore[boto3]==2.5.1 ; python_full_version == "3.9.13"
aiohttp-retry==2.8.3 ; python_full_version == "3.9.13"
aiohttp==3.8.4 ; python_full_version == "3.9.13"
aioitertools==0.11.0 ; python_full_version == "3.9.13"
aiosignal==1.3.1 ; python_full_version == "3.9.13"
amqp==5.1.1 ; python_full_version == "3.9.13"
antlr4-python3-runtime==4.9.3 ; python_full_version == "3.9.13"
anyio==4.4.0 ; python_full_version == "3.9.13"
appdirs==1.4.4 ; python_full_version == "3.9.13"
async-timeout==4.0.2 ; python_full_version == "3.9.13"
asyncssh==2.13.2 ; python_full_version == "3.9.13"
atpublic==4.0 ; python_full_version == "3.9.13"
attrs==23.1.0 ; python_full_version == "3.9.13"
billiard==4.1.0 ; python_full_version == "3.9.13"
boto3==1.26.161 ; python_full_version == "3.9.13"
botocore==1.29.161 ; python_full_version == "3.9.13"
celery==5.3.1 ; python_full_version == "3.9.13"
certifi==2023.5.7 ; python_full_version == "3.9.13"
cffi==1.15.1 ; python_full_version == "3.9.13"
charset-normalizer==3.1.0 ; python_full_version == "3.9.13"
click-didyoumean==0.3.0 ; python_full_version == "3.9.13"
click-plugins==1.1.1 ; python_full_version == "3.9.13"
click-repl==0.3.0 ; python_full_version == "3.9.13"
click==8.1.3 ; python_full_version == "3.9.13"
colorama==0.4.6 ; python_full_version == "3.9.13"
coloredlogs==15.0.1 ; python_full_version == "3.9.13"
configobj==5.0.8 ; python_full_version == "3.9.13"
cryptography==41.0.1 ; python_full_version == "3.9.13"
dependency-injector==4.41.0 ; python_full_version == "3.9.13"
dictdiffer==0.9.0 ; python_full_version == "3.9.13"
dictionary-lib==0.1.38 ; python_full_version == "3.9.13"
diskcache==5.6.1 ; python_full_version == "3.9.13"
distro==1.8.0 ; python_full_version == "3.9.13"
dpath==2.1.6 ; python_full_version == "3.9.13"
dulwich==0.21.5 ; python_full_version == "3.9.13"
dvc-data==2.3.3 ; python_full_version == "3.9.13"
dvc-http==2.30.2 ; python_full_version == "3.9.13"
dvc-objects==0.23.0 ; python_full_version == "3.9.13"
dvc-render==0.5.3 ; python_full_version == "3.9.13"
dvc-s3==2.23.0 ; python_full_version == "3.9.13"
dvc-studio-client==0.11.0 ; python_full_version == "3.9.13"
dvc-task==0.3.0 ; python_full_version == "3.9.13"
dvc==3.4.0 ; python_full_version == "3.9.13"
dynaconf==3.2.6 ; python_full_version == "3.9.13"
exceptiongroup==1.2.2 ; python_full_version == "3.9.13"
fastapi==0.92.0 ; python_full_version == "3.9.13"
filelock==3.12.2 ; python_full_version == "3.9.13"
flatbuffers==24.3.25 ; python_full_version == "3.9.13"
flatten-dict==0.4.2 ; python_full_version == "3.9.13"
flufl-lock==7.1.1 ; python_full_version == "3.9.13"
frozenlist==1.3.3 ; python_full_version == "3.9.13"
fsspec==2023.6.0 ; python_full_version == "3.9.13"
fsspec[http]==2023.6.0 ; python_full_version == "3.9.13"
funcy==2.0 ; python_full_version == "3.9.13"
gitdb==4.0.10 ; python_full_version == "3.9.13"
gitpython==3.1.31 ; python_full_version == "3.9.13"
grandalf==0.8 ; python_full_version == "3.9.13"
h11==0.14.0 ; python_full_version == "3.9.13"
huggingface-hub==0.24.5 ; python_full_version == "3.9.13"
humanfriendly==10.0 ; python_full_version == "3.9.13"
hydra-core==1.3.2 ; python_full_version == "3.9.13"
idna==3.4 ; python_full_version == "3.9.13"
iterative-telemetry==0.0.8 ; python_full_version == "3.9.13"
jmespath==1.0.1 ; python_full_version == "3.9.13"
kombu==5.3.1 ; python_full_version == "3.9.13"
markdown-it-py==3.0.0 ; python_full_version == "3.9.13"
mdurl==0.1.2 ; python_full_version == "3.9.13"
mpmath==1.3.0 ; python_full_version == "3.9.13"
multidict==6.0.4 ; python_full_version == "3.9.13"
nanotime==0.5.2 ; python_full_version == "3.9.13"
networkx==3.1 ; python_full_version == "3.9.13"
numpy==1.25.0 ; python_full_version == "3.9.13"
omegaconf==2.3.0 ; python_full_version == "3.9.13"
onnxruntime==1.18.1 ; python_full_version == "3.9.13"
orjson==3.9.1 ; implementation_name == "cpython" and python_full_version == "3.9.13"
packaging==23.1 ; python_full_version == "3.9.13"
pandas==2.0.3 ; python_full_version == "3.9.13"
pathspec==0.11.1 ; python_full_version == "3.9.13"
platformdirs==3.8.0 ; python_full_version == "3.9.13"
poetry-core==1.9.0 ; python_full_version == "3.9.13"
prompt-toolkit==3.0.39 ; python_full_version == "3.9.13"
protobuf==5.27.3 ; python_full_version == "3.9.13"
psutil==5.9.5 ; python_full_version == "3.9.13"
pyahocorasick==2.0.0 ; python_full_version == "3.9.13"
pyaml==21.10.1 ; python_full_version == "3.9.13"
pycparser==2.21 ; python_full_version == "3.9.13"
pydantic==1.10.17 ; python_full_version == "3.9.13"
pydot==1.4.2 ; python_full_version == "3.9.13"
pygit2==1.12.2 ; python_full_version == "3.9.13"
pygments==2.15.1 ; python_full_version == "3.9.13"
pygtrie==2.5.0 ; python_full_version == "3.9.13"
pyparsing==3.1.0 ; python_full_version == "3.9.13"
pyreadline3==3.4.1 ; sys_platform == "win32" and python_full_version == "3.9.13"
python-dateutil==2.8.2 ; python_full_version == "3.9.13"
pytz==2023.3 ; python_full_version == "3.9.13"
pywin32==306 ; sys_platform == "win32" and python_full_version == "3.9.13"
pyyaml==6.0 ; python_full_version == "3.9.13"
regex==2024.7.24 ; python_full_version == "3.9.13"
requests==2.31.0 ; python_full_version == "3.9.13"
rich==13.4.2 ; python_full_version == "3.9.13"
ruamel-yaml-clib==0.2.7 ; platform_python_implementation == "CPython" and python_full_version == "3.9.13"
ruamel-yaml==0.17.32 ; python_full_version == "3.9.13"
s3fs==2023.6.0 ; python_full_version == "3.9.13"
s3transfer==0.6.1 ; python_full_version == "3.9.13"
safetensors==0.4.4 ; python_full_version == "3.9.13"
scmrepo==1.0.4 ; python_full_version == "3.9.13"
setuptools==68.0.0 ; python_full_version == "3.9.13"
shortuuid==1.0.11 ; python_full_version == "3.9.13"
shtab==1.6.2 ; python_full_version == "3.9.13"
six==1.16.0 ; python_full_version == "3.9.13"
smmap==5.0.0 ; python_full_version == "3.9.13"
sniffio==1.3.1 ; python_full_version == "3.9.13"
sqltrie==0.7.0 ; python_full_version == "3.9.13"
starlette==0.25.0 ; python_full_version == "3.9.13"
sympy==1.13.1 ; python_full_version == "3.9.13"
tabulate==0.9.0 ; python_full_version == "3.9.13"
tokenizers==0.19.1 ; python_full_version == "3.9.13"
tomlkit==0.11.8 ; python_full_version == "3.9.13"
tqdm==4.65.0 ; python_full_version == "3.9.13"
transformers==4.44.0 ; python_full_version == "3.9.13"
typing-extensions==4.7.1 ; python_full_version == "3.9.13"
tzdata==2023.3 ; python_full_version == "3.9.13"
urllib3==1.26.16 ; python_full_version == "3.9.13"
uvicorn==0.20.0 ; python_full_version == "3.9.13"
vine==5.0.0 ; python_full_version == "3.9.13"
voluptuous==0.13.1 ; python_full_version == "3.9.13"
wcwidth==0.2.6 ; python_full_version == "3.9.13"
wrapt==1.15.0 ; python_full_version == "3.9.13"
yarl==1.9.2 ; python_full_version == "3.9.13"
zc-lockfile==3.0.post1 ; python_full_version == "3.9.13"
